+++ src/lib/server/entity.ts
@@ -18,33 +18,6 @@ import {
   ZodSymbol
 } from 'zod';
 
-type DefaultFields<T extends AnyZodObject> = Partial<{
-  [Property in keyof z.infer<T>]:
-    | z.infer<T>[Property]
-    | ((
-        value: z.infer<T>[Property] | null | undefined,
-        data: z.infer<T>
-      ) => z.infer<T>[Property] | null | undefined);
-}>;
-
-function setValidationDefaults<T extends AnyZodObject>(
-  data: z.infer<T>,
-  fields: DefaultFields<T>
-) {
-  for (const stringField of Object.keys(fields)) {
-    const field = stringField as keyof typeof data;
-    const currentData = data[field];
-
-    if (typeof fields[field] === 'function') {
-      // eslint-disable-next-line @typescript-eslint/ban-types
-      const func = fields[field] as Function;
-      data[field] = func(currentData, data);
-    } else if (!currentData) {
-      data[field] = fields[field] as never;
-    }
-  }
-}
-
 export type ZodTypeInfo = {
   zodType: ZodTypeAny;
   isNullable: boolean;
@@ -54,39 +27,42 @@ export type ZodTypeInfo = {
 
 type EntityRecord<T extends AnyZodObject, K> = Record<keyof z.infer<T>, K>;
 
-export type DefaultEntityOptions<T extends AnyZodObject> = {
-  defaults?: DefaultFields<T>;
-  implicitDefaults?: boolean;
-};
-
 //////////////////////////////////////////////////////////////////////////////
 
-type EntityMetaData<T extends AnyZodObject> = {
+export type EntityMetaData<T extends AnyZodObject> = {
   types: EntityRecord<T, string>;
 };
 
-type Entity<T extends AnyZodObject> = {
+export type Entity<T extends AnyZodObject> = {
   typeInfo: EntityRecord<T, ZodTypeInfo>;
   defaultEntity: z.infer<T>;
   constraints: EntityRecord<T, InputConstraints | undefined>;
   meta: EntityMetaData<T>;
 };
 
-export function entityData<T extends AnyZodObject>(
-  schema: T,
-  options: DefaultEntityOptions<T> = {}
-) {
-  if (entityCache.has(schema)) return entityCache.get(schema) as Entity<T>;
+export function entityData<T extends AnyZodObject>(schema: T) {
+  const cached = getCached(schema);
+  if (cached) return cached;
 
   const typeInfos = typeInfo(schema);
+  const defaultEnt = defaultEntity(schema);
   const entity: Entity<T> = {
     typeInfo: typeInfos,
-    defaultEntity: defaultEntity(schema, options),
+    defaultEntity: defaultEnt,
     constraints: constraints(schema, typeInfos),
     meta: meta(schema)
   };
-  entityCache.set(schema, entity);
+
+  setCached(schema, entity);
   return entity;
+}
+
+function setCached<T extends AnyZodObject>(schema: T, entity: Entity<T>) {
+  entityCache.set(schema, entity);
+}
+
+function getCached<T extends AnyZodObject>(schema: T) {
+  return entityCache.get(schema) as Entity<T>;
 }
 
 const entityCache: WeakMap<AnyZodObject, Entity<AnyZodObject>> = new WeakMap<
@@ -133,73 +109,83 @@ function typeInfo<T extends AnyZodObject>(schema: T) {
   return _mapSchema(schema, (obj) => zodTypeInfo(obj));
 }
 
+export function checkMissingFields<T extends AnyZodObject>(
+  schema: T,
+  data: Partial<z.infer<T>> | null | undefined
+) {
+  const entity = entityData(schema);
+
+  console.log(data, entity.constraints);
+
+  const missingFields = Object.keys(entity.constraints).filter(
+    (field) =>
+      entity.constraints[field]?.required &&
+      entity.defaultEntity[field] === undefined &&
+      (!data || data[field] === undefined || data[field] === null)
+  );
+
+  if (missingFields.length) {
+    const errors = missingFields.map(
+      (field) =>
+        `"${String(field)}" (${
+          entity.typeInfo[field].zodType.constructor.name
+        })`
+    );
+    throw new Error(
+      `Unsupported default value for schema field(s): ${errors.join(
+        ', '
+      )}. Add default, optional or nullable to those fields in the schema.`
+    );
+  }
+}
+
+export function valueOrDefault(
+  value: unknown,
+  strict: boolean,
+  implicitDefaults: true,
+  typeInfo: ZodTypeInfo
+) {
+  if (value) return value;
+
+  const { zodType, isNullable, isOptional, defaultValue } = typeInfo;
+
+  // Based on schema type, check what the empty value should be parsed to
+
+  // For convenience, make undefined into nullable if possible.
+  // otherwise all nullable fields requires a default value or optional.
+  // In the database, null is assumed if no other value (undefined doesn't exist there),
+  // so this should be ok.
+  // Also make a check for strict, so empty strings from FormData can also be set here.
+
+  if (strict && value !== undefined) return value;
+  if (defaultValue !== undefined) return defaultValue;
+  if (isNullable) return null;
+  if (isOptional) return undefined;
+
+  if (implicitDefaults) {
+    if (zodType instanceof ZodString) return '';
+    if (zodType instanceof ZodNumber) return 0;
+    if (zodType instanceof ZodBoolean) return false;
+    if (zodType instanceof ZodArray) return [];
+    if (zodType instanceof ZodObject) return {};
+    if (zodType instanceof ZodBigInt) return BigInt(0);
+    if (zodType instanceof ZodSymbol) return Symbol();
+  }
+
+  return undefined;
+}
+
 /**
  * Returns the default values for a zod validation schema.
  * The main gotcha is that undefined values are changed to null if the field is nullable.
  */
 export function defaultEntity<T extends AnyZodObject>(
-  schema: T,
-  options: {
-    defaults?: DefaultFields<T>;
-    implicitDefaults?: boolean;
-  } = {}
+  schema: T
 ): z.infer<T> {
-  function valueOrDefault<T extends AnyZodObject>(
-    field: keyof z.infer<T>,
-    value: unknown,
-    strict: boolean,
-    implicitDefaults: boolean,
-    typeInfo: ZodTypeInfo
-  ) {
-    if (value) return value;
-
-    const { zodType, isNullable, isOptional, defaultValue } = typeInfo;
-
-    // Based on schema type, check what the empty value should be parsed to
-    function emptyValue() {
-      // For convenience, make undefined into nullable if possible.
-      // otherwise all nullable fields requires a default value or optional.
-      // In the database, null is assumed if no other value (undefined doesn't exist there),
-      // so this should be ok.
-      // Also make a check for strict, so empty strings from FormData can also be set here.
-      if (strict && value !== undefined) return value;
-      if (defaultValue !== undefined) return defaultValue;
-      if (isNullable) return null;
-      if (isOptional) return undefined;
-
-      if (implicitDefaults) {
-        if (zodType instanceof ZodString) return '';
-        if (zodType instanceof ZodNumber) return 0;
-        if (zodType instanceof ZodBoolean) return false;
-        if (zodType instanceof ZodArray) return [];
-        if (zodType instanceof ZodObject) return {};
-        if (zodType instanceof ZodBigInt) return BigInt(0);
-        if (zodType instanceof ZodSymbol) return Symbol();
-      }
-
-      throw new Error(
-        `Unsupported type for ${strict ? 'strict' : 'falsy'} ${
-          implicitDefaults ? 'implicit' : 'explicit'
-        } values on field "${String(field)}": ${
-          zodType.constructor.name
-        }. Add default, optional or nullable to the schema.`
-      );
-    }
-
-    return emptyValue();
-  }
-
-  options = { ...options };
-
   const fields = Object.keys(schema.keyof().Values);
 
   let output: Record<string, unknown> = {};
   let defaultKeys: string[] | undefined;
-
-  if (options.defaults) {
-    setValidationDefaults(output, options.defaults);
-    defaultKeys = Object.keys(options.defaults);
-  }
 
   const schemaTypeInfo = typeInfo(schema);
 
@@ -207,18 +193,12 @@ export function defaultEntity<T extends AnyZodObject>(
   output = Object.fromEntries(
     fields.map((field) => {
       const typeInfo = schemaTypeInfo[field];
-      const value =
+      const newValue =
         defaultKeys && defaultKeys.includes(field)
           ? output[field]
-          : valueOrDefault(
-              field,
-              undefined,
-              true,
-              options.implicitDefaults ?? true,
-              typeInfo
-            );
+          : valueOrDefault(undefined, true, true, typeInfo);
 
-      return [field, value];
+      return [field, newValue];
     })
   );
 
@@ -278,7 +258,13 @@ function constraints<T extends AnyZodObject>(
       if (zodType.maxDate) output.max = zodType.maxDate.toISOString();
     }
 
-    if (!info.isNullable && !info.isOptional) output.required = true;
+    if (
+      !info.isNullable &&
+      !info.isOptional &&
+      (info.defaultValue === undefined || info.defaultValue === null)
+    ) {
+      output.required = true;
+    }
 
     return Object.keys(output).length > 0 ? output : undefined;
   }