+++ CHANGELOG.md
@@ -5,6 +5,14 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.16] - 2023-03-01
+
+### Fixed
+
+- Checking for missing fields only if no `FormData` exists.
+- Correct comparison for client-side `Date` validation.
+- Improved cacheability for validation data
+
 ## [0.5.15] - 2023-02-28
 
 ### Added